requirements.txt
flask==3.0.2
openai>=1.12.0
python-dotenv==1.0.1
requests==2.31.0
pydantic>=2.0.0
typing-extensions>=4.5.0
agentops==0.4.3
nest_asyncio==1.6.0
colorama==0.4.6